tlib update and rockspec removal.
This update adds functions to handle command line arguments, such as:

	* check_args
	* verify_args
	* get_arg

and the most interesting function in my opinion, the function to import
modules using relative and absolute paths.

I intend to create more functions to manage modules, for now we have:

	* import
	* check_lib

For the next update I hope to improve the existing functions and bring more
for handling files and directories.